Commit - after S10

+++ RestaurantAPI/Services/RestaurantService.cs
@@ -1,15 +1,19 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using RestaurantAPI.Authorization;
 using RestaurantAPI.Entities;
 using RestaurantAPI.Exceptions;
 using RestaurantAPI.Models;
+using System.Linq.Expressions;
+using System.Security.Claims;
 
 namespace RestaurantAPI.Services
 {
     public interface IRestaurantService
     {
         int Create(CreateRestaurantDto dto);
-        IEnumerable<RestaurantDto> GetAll();
+        PageResult<RestaurantDto> GetAll(RestaurantQuery query);
         RestaurantDto GetById(int id);
         public void Delete(int id);
         public void Modify(int id, UpdateRestaurantDto dto);
@@ -19,12 +23,16 @@ namespace RestaurantAPI.Services
 
     public class RestaurantService : IRestaurantService
     {
+        private readonly IUserContextService _userContextService;
+        private readonly IAuthorizationService _authorizationService;
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
 
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger, IAuthorizationService authorizationService, IUserContextService userContextService)
         {
+            _userContextService = userContextService;
+            _authorizationService = authorizationService;
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
@@ -45,30 +53,49 @@ namespace RestaurantAPI.Services
             return result;
         }
 
-        public IEnumerable<RestaurantDto> GetAll()
+        public PageResult<RestaurantDto> GetAll(RestaurantQuery query)
         {
-            var restaurants = _dbContext
+            var baseQuery = _dbContext
                 .Restaurants
                 .Include(r => r.Address)
                 .Include(r => r.Dishes)
-                .ToList();
+                .Where(r => query.SearchPhrase == null ||
+                    r.Name.ToLower().Contains(query.SearchPhrase.ToLower()) ||
+                    r.Description.ToLower().Contains(query.SearchPhrase.ToLower()));
 
-            //var restaurantsDtos = restaurants.Select(r => new RestaurantDto()
-            //{
-            //    Id = r.Id,
-            //    Name = r.Name,
-            //    Category = r.Category,
-            //    ...
-            //})
+            if(!string.IsNullOrEmpty(query.SortBy))
+            {
+                var columnsSelectors = new Dictionary<string, Expression<Func<Restaurant, object>>>
+                {
+                    { nameof(Restaurant.Name), r => r.Name},
+                    { nameof(Restaurant.Description), r => r.Description},
+                    { nameof(Restaurant.Category), r => r.Category},
+
+                };
+
+                var selectedColumn = columnsSelectors[query.SortBy];
+
+                //baseQuery.OrderBy(r => r.Name);
+                baseQuery = query.SortDirection == SortDirection.ASC ? 
+                    baseQuery.OrderBy( r => r.Name) : baseQuery.OrderByDescending( r => r.Name);
+            }
+
+            var restaurants = baseQuery
+                .Skip(query.PageSize * (query.PageNumber - 1))
+                .Take(query.PageSize)
+                .ToList();
 
             var restaurantsDtos = _mapper.Map<List<RestaurantDto>>(restaurants);
 
-            return restaurantsDtos;
+            var result = new PageResult<RestaurantDto>(restaurantsDtos, baseQuery.Count(), query.PageSize, query.PageNumber);
+
+            return result;
         }
 
         public int Create(CreateRestaurantDto dto)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = _userContextService.GetUserId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
 
@@ -86,6 +113,13 @@ namespace RestaurantAPI.Services
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found.");
 
+            var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
+
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
 
@@ -93,13 +127,21 @@ namespace RestaurantAPI.Services
 
         public void Modify(int id, UpdateRestaurantDto dto)
         {
+
             var restaurant = _dbContext
                 .Restaurants
                 .FirstOrDefault(r => r.Id == id);
-            // do id dodać 3 paramsy albo jakis obiekt, potem zmienic pola restaurant dać 
+ 
 
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found.");
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             restaurant.Name = dto.Name;
             restaurant.Description = dto.Description;
